screw-api/requirements.txt
sagemaker==1.9.0
numpy==1.15.1
Pillow==5.2.0